+++ app/src/androidTest/java/net/cryptodirect/authenticator/LinkAccountBehaviorTest.java
@@ -91,7 +91,7 @@ public class LinkAccountBehaviorTest
         // The current TOTP code and time-wheel should be displayed for the newly linked account
         onView(withId(R.id.code_box)).check(matches(isDisplayed()));
         onView(withId(R.id.code_box)).check(matches(withText(generateTOTP(
-                Base64.getDecoder().decode(key), (long) getTC(30), 6, SHA256))));
+                Base64.getDecoder().decode(key), getTC(30), 6, SHA256))));
     }
 
     @Test
@@ -136,7 +136,7 @@ public class LinkAccountBehaviorTest
         // The current TOTP code and time-wheel should be displayed for the newly linked account
         onView(withId(R.id.code_box)).check(matches(isDisplayed()));
         onView(withId(R.id.code_box)).check(matches(withText(generateTOTP(
-                Base64.getDecoder().decode(key), (long) getTC(30), 6, SHA256))));
+                Base64.getDecoder().decode(key), getTC(30), 6, SHA256))));
     }
 
     @Test
@@ -177,7 +177,7 @@ public class LinkAccountBehaviorTest
         // The current TOTP code and time-wheel should be displayed for the newly linked account
         onView(withId(R.id.code_box)).check(matches(isDisplayed()));
         onView(withId(R.id.code_box)).check(matches(withText(generateTOTP(
-                Base32.getDecoder().decode(key), (long) getTC(30), 6, SHA1))));
+                Base32.getDecoder().decode(key), getTC(30), 6, SHA1))));
     }
 
     @Test
@@ -217,6 +217,6 @@ public class LinkAccountBehaviorTest
         // The current TOTP code and time-wheel should be displayed for the newly linked account
         onView(withId(R.id.code_box)).check(matches(isDisplayed()));
         onView(withId(R.id.code_box)).check(matches(withText(generateTOTP(
-                Base32.getDecoder().decode(key), (long) getTC(30), 6, SHA1))));
+                Base32.getDecoder().decode(key), getTC(30), 6, SHA1))));
     }
 }
